+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 ## Breaking Changes:
 
 - Enable dataset augmentation customizing by `@illian01` in [PR 201](https://github.com/Nota-NetsPresso/netspresso-trainer/pull/201)
+- Add postprocessor module by `@illian01` in [PR 223](https://github.com/Nota-NetsPresso/netspresso-trainer/pull/223)
 
 ## Other Changes: